pubsubfirestore/requirements.txt
google-cloud-firestore==2.22.0
functions-framework==3.5.0
